Reorganize SQL to Slack Operators
